BankApp: HTMLception finished

+++ week-07/day-01/src/main/java/bank/bankapp/models/BankAccount.java
@@ -2,10 +2,11 @@ package bank.bankapp.models;
 
 public class BankAccount {
     private String name;
-    private int balance;
+    private double balance;
     private String animalType;
 
-    public BankAccount(String name, int balance, String animalType) {
+
+    public BankAccount(String name, double balance, String animalType) {
         this.name = name;
         this.balance = balance;
         this.animalType = animalType;
@@ -22,11 +23,11 @@ public class BankAccount {
         this.name = name;
     }
 
-    public int getBalance() {
+    public double getBalance() {
         return balance;
     }
 
-    public void setBalance(int balance) {
+    public void setBalance(double balance) {
         this.balance = balance;
     }
